Changed access method to EpetraVector.
The access to EpetraVectors are now done by using the
SumIntoGlobalValues method instead of the bracket (operator[]). We're
doing this because the IDs of the volumes are global and the bracket
operator takes local IDs and maps into global ones, causing
inconsistence in the final value stored in the vector.